requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
argon2-cffi==19.2.0
asgiref==3.1.2            # via daphne
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via asgiref
attrs==19.1.0             # via automat, twisted
autobahn==19.6.2          # via daphne
automat==0.7.0            # via twisted
certifi==2019.6.16        # via requests
cffi==1.12.3              # via argon2-cffi, cryptography
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
constantly==15.1.0        # via twisted
cryptography==2.7         # via autobahn
daphne==2.3.0
django-allow-cidr==0.3.1
django-celery-beat==1.5.0
django-cors-headers==3.0.2
django-mysql==3.2.0
django-redis==4.10.0
django-rest-camel==0.3.4
django-timezone-field==3.0  # via django-celery-beat
django==2.2.2
djangorestframework==3.9.4
gunicorn==20.0.0
hyperlink==19.0.0         # via twisted
idna==2.8                 # via hyperlink, requests
incremental==17.5.0       # via twisted
invoke==1.2.0
mysqlclient==1.4.2.post1
netaddr==0.7.19           # via django-allow-cidr
pip-tools==3.8.0
pycparser==2.19           # via cffi
pyhamcrest==1.9.0         # via twisted
python-crontab==2.3.8     # via django-celery-beat
python-dateutil==2.8.0    # via python-crontab
pytz==2019.1              # via django, django-timezone-field
redis==3.2.1              # via django-redis
requests==2.22.0
six==1.12.0               # via argon2-cffi, autobahn, automat, cryptography, pip-tools, pyhamcrest, python-dateutil, txaio
sqlparse==0.3.0           # via django
twisted==19.2.1           # via daphne
txaio==18.8.1             # via autobahn
urllib3==1.25.3           # via requests
zope.interface==4.6.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.2        # via gunicorn, pyhamcrest, zope.interface
